got loading functionality to work
