+++ src/components/mainPanel/leftPanel/LeftPanel.svelte
@@ -1,12 +1,21 @@
 <script lang="ts">
   import { onMount } from 'svelte';
-  // let timeOver = false;
-
-  // setTimeout(() => (timeOver = !timeOver), 10000);
+  import { rootNodes } from '../../../store';
+  let timeOver = false;
+  import DummyTree2 from './dummyTree2.svelte';
+  setTimeout(() => (timeOver = !timeOver), 10000);
+  rootNodes.subscribe(() => console.log($rootNodes, Date.now()));
 </script>
 
-<!-- {#if timeOver}{/if} -->
-<div class="container" />
+<!-- check to see if rootNodes is undefined or not -->
+
+{#if $rootNodes.length}
+  <div class="container">
+    <DummyTree2 />
+  </div>
+{:else}
+  <div class="container">not yet ready</div>
+{/if}
 
 <style>
   .container {
